Add files via upload
This code contains system identification methods to deduce the transfer function used in the BOSE noise cancellation feedback algorithm. 
The input signal is the voltage to internal ear-cup s[eaker (random white noise with variance 1.1).
The output signal is an amplified voltage signal of the internal ear-cup microphone.
Parametric and non-parametric methods are used to predict the transfer function of the model.